+++ platform/core/ffi/export/ios_headers/polypod-core.h
@@ -13,38 +13,9 @@ typedef struct BridgeToPlatform {
   struct CByteBuffer (*perform_request)(struct CByteBuffer request);
 } BridgeToPlatform;
 
-/**
- * # Safety
- * This function can be unsafe if the language_code pointer is null or the string is in wrong format.
- *
- * Mention - It is needed to be tested in integration if `*const u8` is the appropriate return format.
- *           Also, most likely, it will be required to expose an API to deallocate the byte buffer after parsing.
- *
- * Bootstrap core with the given configuration:
- * - language_code: User's locale language code.
- * Returns a MessagePack byte array with core_bootstrap_response.
- */
 struct CByteBuffer core_bootstrap(struct CByteBuffer args,
                                   struct BridgeToPlatform bridge);
 
 struct CByteBuffer execute_request(struct CByteBuffer core_request);
 
-/**
- * Executes the given RDF query.
- * Returns Result<String, CoreFailure> as MessagePack value.
- */
-struct CByteBuffer exec_rdf_query(const char *query);
-
-/**
- * Executes the given RDF update.
- * Returns Result<Void, CoreFailure> as MessagePack value.
- */
-struct CByteBuffer exec_rdf_update(const char *update);
-
-/**
- * # Safety
- * This function can be unsafe if trying to deallocate invalid memory.
- *
- * Drops the given bytes.
- */
 void free_bytes(uint8_t *bytes);
